move Hypre periodic BC test
This test seems to be messing up the code coverage checks since the fortran needs recompiling to run it. Hopefully this fixes the issue.
